Moved replaceShorthandObjectMethod to its own file, added comments, and massaged error handling.
